rename IAzResourceTypeLoader to IResourceTypeLoader
